requirements.txt
streamlit
google-genai
python-dotenv
requests
beautifulsoup4
google-search-results
